requirements.txt
pymongo>=4.6.0
python-dotenv>=1.0.0
matplotlib>=3.5.0
flask>=2.0.0
